implementado token

+++ Inventarios/Controllers/TablasMaestras/ActividadesEconomicasController.cs
@@ -1,7 +1,7 @@
 ﻿using Inventarios.DTO.TablasMaestras;
 using Inventarios.Models.TablasMaestras;
 using Inventarios.services.TablasMaestras;
-
+using Inventarios.Utils;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Inventarios.Controllers.TablasMaestras
@@ -13,50 +13,56 @@ namespace Inventarios.Controllers.TablasMaestras
         private readonly ActividadesEconomicasService _service;
 
         private List<ActividadesEconomicasDTO>? list;
+        private Validaciones _validaciones;
 
-        public ActividadesEconomicasController(ActividadesEconomicasService service)
+
+        public ActividadesEconomicasController(ActividadesEconomicasService service, Validaciones validaciones)
         {
             _service = service;
+            _validaciones = validaciones;   
         }
 
         [HttpPost]
         [ActionName("Add")]
-        public Mensaje Add(ActividadesEconomicas obj)
+        public Mensaje Add(ActividadesEconomicas obj, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
             Mensaje list = _service.Add(obj);
             return list;
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id}/{idusuario}/{token}")]
         [ActionName("Delete")]
-        public Mensaje Delete(int id)
+        public Mensaje Delete(int id, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
             Mensaje list = _service.Delete(id);
             return list;
         }
 
         [HttpPut]
         [ActionName("Update")]
-        public Mensaje Update(ActividadesEconomicas obj)
+        public Mensaje Update(ActividadesEconomicas obj, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
             Mensaje list = _service.Update(obj);
             return list;
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id}/{idusuario}/{token}")]
         [ActionName("GetById")]
-        public List<ActividadesEconomicas>? GetById(int id)
+        public List<ActividadesEconomicas>? GetById(int id, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
             List<ActividadesEconomicas> list = _service.GetById(id);
             return list;
         }
 
-        [HttpGet("{filtro}")]
+        [HttpGet("{filtro}/{idusuario}/{token}")]
         [ActionName("GetAll")]
-        public List<ActividadesEconomicasDTO>? GetAll(string filtro = "")
+        public List<ActividadesEconomicasDTO>? GetAll(string filtro, int idusuario, string token)
         {
-            if (filtro == "undefined") filtro = "";
-
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
             list = _service.List(filtro);
             return list;
         }
