+++ Inventarios/Controllers/TablasMaestras/ColoresController.cs
@@ -1,7 +1,7 @@
 ﻿using Inventarios.DTO.TablasMaestras;
 using Inventarios.Models.TablasMaestras;
 using Inventarios.services.TablasMaestras;
-
+using Inventarios.Utils;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Inventarios.Controllers.TablasMaestras
@@ -11,52 +11,59 @@ namespace Inventarios.Controllers.TablasMaestras
     public class ColoresController : ControllerBase
     {
         private readonly ColoresService _service;
+        private Validaciones _validaciones;
 
         private List<ColoresDTO>? list;
 
-        public ColoresController(ColoresService service)
+
+        public ColoresController(ColoresService service, Validaciones validaciones)
         {
             _service = service;
+            _validaciones = validaciones;   
         }
 
         [HttpPost]
         [ActionName("Add")]
-        public Mensaje Add(Colores obj)
+        public Mensaje Add(Colores obj, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
             Mensaje list = _service.Add(obj);
             return list;
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id}/{idusuario}/{token}")]
         [ActionName("Delete")]
-        public Mensaje Delete(int id)
+        public Mensaje Delete(int id, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
             Mensaje list = _service.Delete(id);
             return list;
         }
 
         [HttpPut]
         [ActionName("Update")]
-        public Mensaje Update(Colores obj)
+        public Mensaje Update(Colores obj, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
             Mensaje list = _service.Update(obj);
             return list;
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id}/{idusuario}/{token}")]
         [ActionName("GetById")]
-        public List<Colores>? GetById(int id)
+        public List<Colores>? GetById(int id, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
             List<Colores> list = _service.GetById(id);
             return list;
         }
 
-        [HttpGet("{filtro}")]
+        [HttpGet("{filtro}/{idusuario}/{token}")]
         [ActionName("GetAll")]
-        public List<ColoresDTO>? GetAll(string filtro = "")
+        public List<ColoresDTO>? GetAll(string filtro , int idusuario, string token)
         {
-            if (filtro == "undefined") filtro = "";
 
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
             list = _service.List(filtro);
             return list;
         }
