+++ Inventarios/Controllers/TablasMaestras/EstadosDeUnRegistroController.cs
@@ -1,7 +1,7 @@
 ﻿using Inventarios.DTO.TablasMaestras;
 using Inventarios.Models.TablasMaestras;
 using Inventarios.services.TablasMaestras;
-
+using Inventarios.Utils;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Inventarios.Controllers.TablasMaestras
@@ -11,52 +11,57 @@ namespace Inventarios.Controllers.TablasMaestras
     public class EstadosDeUnRegistroController : ControllerBase
     {
         private readonly EstadosDeUnRegistroService _service;
+        private Validaciones _validaciones;
 
         private List<EstadosDeUnRegistroDTO>? list;
 
-        public EstadosDeUnRegistroController(EstadosDeUnRegistroService service)
+        public EstadosDeUnRegistroController(EstadosDeUnRegistroService service, Validaciones validaciones)
         {
             _service = service;
+            _validaciones = validaciones;
         }
 
         [HttpPost]
         [ActionName("Add")]
-        public Mensaje Add(EstadosDeUnRegistro obj)
+        public Mensaje Add(EstadosDeUnRegistro obj, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return new Mensaje();
             Mensaje list = _service.Add(obj);
             return list;
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id}/{idusuario}/{token}")]
         [ActionName("Delete")]
-        public Mensaje Delete(int id)
+        public Mensaje Delete(int id, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return new Mensaje();
             Mensaje list = _service.Delete(id);
             return list;
         }
 
         [HttpPut]
         [ActionName("Update")]
-        public Mensaje Update(EstadosDeUnRegistro obj)
+        public Mensaje Update(EstadosDeUnRegistro obj, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return new Mensaje();
             Mensaje list = _service.Update(obj);
             return list;
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id}/{idusuario}/{token}")]
         [ActionName("GetById")]
-        public List<EstadosDeUnRegistro>? GetById(int id)
+        public List<EstadosDeUnRegistro>? GetById(int id, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
             List<EstadosDeUnRegistro> list = _service.GetById(id);
             return list;
         }
 
-        [HttpGet("{filtro}")]
+        [HttpGet("{filtro}/{idusuario}/{token}")]
         [ActionName("GetAll")]
-        public List<EstadosDeUnRegistroDTO>? GetAll(string filtro = "")
+        public List<EstadosDeUnRegistroDTO>? GetAll(string filtro, int idusuario, string token)
         {
-            if (filtro == "undefined") filtro = "";
-
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
             list = _service.List(filtro);
             return list;
         }
